+++ chaosaws/ssm/actions.py
@@ -7,7 +7,7 @@ from chaoslib.types import Configuration, Secrets
 from chaosaws import aws_client
 from chaosaws.types import AWSResponse
 
-__all__ = ["create_document", "send_command", "delete_document"]
+__all__ = ["create_document", "send_command", "delete_document", "put_parameter"]
 
 
 def create_document(
@@ -125,3 +125,84 @@ def delete_document(
                 name, str(e.response["Error"]["Message"])
             )
         )
+
+
+def put_parameter(
+    name: str,
+    value: str,
+    description: str = None,
+    type: str = None,
+    key_id: str = None,
+    overwrite: bool = False,
+    allowed_pattern: str = None,
+    tags: List[Dict[str, str]] = None,
+    tier: str = None,
+    policies: str = None,
+    data_type: str = None,
+    configuration: Configuration = None,
+    secrets: Secrets = None,
+) -> AWSResponse:
+    """
+    Add or update a parameter in the Systems Manager Parameter Store.
+
+    :param name: str name of the parameter
+    :param value: str value of the parameter
+    :param description: str information about the parameter
+    :param type: str type of the paramater value, such as 'String'
+    :param key_id: str KMS key id to use while encrypting the parameter value
+    :param overwrite: bool allow the parameter value to be overwritten
+    :param allowed_pattern: str regex to validate parameter value
+    :param tags: List[Dict[str, str]] metadata about the parameter
+    :param tier: str storage classes such as 'Advanced' to allow larger parameter
+        values
+    :param policies: str storage policies such as expiration in JSON format
+    :param data_type: str data type for String. Allows the validation of AMI IDs
+    :param configuration: Configuration object representing the CTK Configuration
+    :param secrets: Secret object representing the CTK Secrets
+    :returns: dict representing the Version and Tier of the parameter
+
+    Examples
+    --------
+    >>> Configuration within experiment
+       {
+            "name": "Activate Chaos",
+            "type": "action",
+            "provider": {
+                "type": "python",
+                "module": "chaosaws.ssm.actions",
+                "func": "put_parameter",
+                "arguments": {
+                    "name": "chaos_trigger",
+                    "value": true,
+                    "overwrite": true,
+                    "type": "SecureString",
+                }
+            },
+        }
+
+    """
+    if not name:
+        raise ActivityFailed("To create a parameter, you must specify the name")
+
+    if not value:
+        raise ActivityFailed("To create a parameter, you must specify the value")
+
+    try:
+        client = aws_client("ssm", configuration, secrets)
+        kwargs = {
+            "Name": name,
+            "Description": description,
+            "Value": value,
+            "Type": type,
+            "KeyId": key_id,
+            "Overwrite": overwrite,
+            "AllowedPattern": allowed_pattern,
+            "Tags": tags,
+            "Tier": tier,
+            "Policies": policies,
+            "DataType": data_type,
+        }
+        return client.put_parameter(**kwargs)
+    except ClientError as e:
+        error_message = e.response["Error"]["Message"]
+        raise ActivityFailed(f"Failed to put  parameter '{name}': '{error_message}'")